parser/htmlparser/tests/mochitest/html5lib_tree_construction/pending-spec-changes.dat

#data
<input type="hidden"><frameset>
#errors
21: Start tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
31: “frameset” start tag seen.
31: End of file seen and there were open elements.
#document
| <html>
|   <head>
|   <frameset>
